Updated Edit, Delete, and Search
Added error handling and correct sql variable representation to the APIs

+++ api/DeleteContact.php
@@ -3,12 +3,39 @@
 	
 	$inData = getRequestInfo();
 
-	$contactID = $inData["ContactID"];
+	$userID = $inData["UserID"];
+	$ID = $inData["ID"];
 
-	$stmt = $conn->prepare("DELETE FROM Contacts WHERE ID = ?");
-	$stmt->bind_param("i", $contactID);
+	$stmt = $conn->prepare("SELECT UserID, ID FROM Contacts WHERE UserID = ?")
+
+	$stmt->bind_param("i", $userID);
 	$stmt->execute();
+
+	$result = $stmt->get_result();
+
+	if($result->num_rows == 0)
+	{
+		returnWithError("Contact to delete doesn't exist");
+		$stmt->close();
+		$conn->close();
+		exit();
+	}
+
+	$stmt->close();
+	
+	$stmt = $conn->prepare("DELETE FROM Contacts WHERE UserID = ? AND ID = ?");
+	$stmt->bind_param("ii", $userID, $ID);
+	$stmt->execute();
+
+	if($stmt->affected_rows > 0)
+	{
+		returnWithError("Contact successfully deleted");
+	}
+	else
+	{
+		returnWithError("Failed to delete contact");
+	}
+
 	$stmt->close();
 	$conn->close();
-	returnWithError("");
 ?>
